+++ lua/obsidian/backlinks.lua
@@ -98,39 +98,37 @@ Backlinks._gather = function(self)
     return false
   end
 
+  local opts = search.SearchOpts.from_tbl {
+    fixed_strings = true,
+    sort_by = self.client.opts.sort_by,
+    sort_reversed = self.client.opts.sort_reversed,
+  }
+
   local backlink_matches = {}
   local last_path = nil
   local last_note = nil
   local tx, rx = channel.oneshot()
 
-  search.search_async(
-    self.client.dir,
-    "[[" .. tostring(self.note.id),
-    self.client.opts.sort_by,
-    self.client.opts.sort_reversed,
-    { "--fixed-strings" },
-    function(match)
-      if is_valid_backlink(match) then
-        local path = match.path.text
-        local src_note
-        if path ~= last_path then
-          src_note = Note.from_file(path, self.client.dir)
-        else
-          assert(last_note ~= nil)
-          src_note = last_note
-        end
-        table.insert(
-          backlink_matches,
-          { note = src_note, line = match.line_number, text = string.gsub(match.lines.text, "\n", "") }
-        )
-        last_path = path
-        last_note = src_note
+  search.search_async(self.client.dir, "[[" .. tostring(self.note.id), opts, function(match)
+    if is_valid_backlink(match) then
+      local path = match.path.text
+      local src_note
+      if path ~= last_path then
+        src_note = Note.from_file(path, self.client.dir)
+      else
+        assert(last_note ~= nil)
+        src_note = last_note
       end
-    end,
-    function(_, _, _)
-      tx()
+      table.insert(
+        backlink_matches,
+        { note = src_note, line = match.line_number, text = string.gsub(match.lines.text, "\n", "") }
+      )
+      last_path = path
+      last_note = src_note
     end
-  )
+  end, function(_, _, _)
+    tx()
+  end)
 
   rx()